==> src/01-motivation.Rmd <==
# Motivation

## Introduction

Geothermal power makes up 13% of New Zealand's electricity supply, going all the way back to 1958. The Wairakei field's current operator, Contact Energy Ltd., gathers a huge amount of data for the purposes of monitoring the state of the field.

Traditional data management tools (SCADA) give point measurements, which are fed into an Excel model of the steamfield. However, all measurements and models have errors. By ignoring uncertainty, we make ourselves oblivious to the risks posed by randomness and cannot quantify our own uncertainty in our estimates.

## Background

At Wairakei, there is a unique and complex network of facilities that make it possible to generate power under difficult conditions. Wells draw mixed-phase fluid from the reservoir to the surface, where it is separated by flash plants. The resulting steam and water is sent to generators, reinjected or discarded.

Not every flow, temperature or enthalpy can be measured everywhere in the network. Some parameters are calculated, and then used to estimate the state of the network. Live, accurate knowledge of the steamfield is vital for operational decisions such as when to do maintenance, or how to ensure the longevity of its natural resources.

## What are we doing?

[Bayesian statistics](https://en.wikipedia.org/wiki/Bayesian_inference) and [Monte Carlo simulation](https://en.wikipedia.org/wiki/Monte_Carlo_method) are two methods that go hand in hand to incorporate uncertainty in our predictions.

Our implementation in R provides graphical interpretations of the model's outputs. The model is fully configurable using an Excel tool, and automatic diagnostic tests verify that it is operating correctly.

## Outcomes

We iterated through a variety of model formulations and methods, settling on the one with the best performance. Our program also generates interactive plots of the results, allowing the operator to filter by facility and compare the performance of their assets.

We also evaluate the risk of the steamfield exceeding any constraints, such as steam flow limits.

- **Quick Runtime** Execution of data ingestion, simulation, and outputs in less than five minutes. The current workflow takes hours.
- **Intuitive Results** Density plots show our estimates and uncertainties. Risks of constraint violations are also easy to interpret.

### Results

- **Regplot** Our model predicting fluid production for each well incorporates confidence bands.
- **Lintsplot** We can make production forecasts into the future, with an associated confidence interval.
- **Limitplot** We simulated an actual Contact Energy configuration. The steam flows are  unlikely to exceed the flash plant limits, shown as the red lines. This confirms that their configuration is not risky.

## Conclusions

We have developed an efficient model that can quantify uncertainty throughout the network. It is also effective at evaluating the risk associated with each network configuration.

As with all data-driven models, its performance is limited by the quality of the data so we would love to work with Contact to integrate more sources of information.
